- Working Reflection, moving towards SOLID - Build fixed
